+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - option to make attractor repel particles
 - added texture to differentiated body
 - improved "body properties" page, added mass and average density
+- added upper limit of smoothing lengths
 
 ## 2022-01-07
 - attractors can absorb particles